requirements.txt
google-genai==1.2.0
openai==2.31.0
anthropic==0.93.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-mock>=3.12.0
audible==0.10.0
audible-cli==0.3.3
httpx>=0.23.3,<0.28.0
markdownify>=0.13.1
beautifulsoup4>=4.12.0
trafilatura==2.0.0
lxml_html_clean>=0.4.0
pydantic>=2.0
PyYAML>=6.0
